add option to edit simple or complex blog entry
